Get rid of Relm again
It is kind of nice so far, but it is also a bit magic, partially because
it is macro-heavy, and I think I should be able to achieve the same with
just my state an an Arc and scheduling closures on the main loop. This
is not criticism of Relm, it's just that I think for my use case, the
benefits do not outweigh the cost.

Revert "Update window title after dropping file"

This reverts commit bd26a22b22670bc95ce0a16ed6825a3fb0a204b4.

Revert "Use Relm for UI"

This reverts commit 5cb7f7bd1b3c3f0806e3354b14f084366d95da39.