Add custom drawing Cairo widget
Apparently this is the way to go to draw DPI-dependent graphics. While
GtkImage scales up the bitmap -- which is great to run HiDPI-unaware
apps on a HiDPI display -- there is no way to say "I am DPI aware, and I
am able to render more pixels to produce a sharper picture". With Cairo
this is possible, so Cairo is the way to go. That also makes it nice to
overlay a grid and to add labels.

Also, I could render a GdkPixbuf in the background thread, and then send
it over to the UI thread to be drawn.