Intersect windows and pixels to render spectrogram
This ensures proper rendering both in the regime where windows are wider
than a pixel, and when they are narrower than a pixel.

The only thing that is not right now, is that windows have a uniform
weight (proportional to area), whereas for the spectrogram, we use the
Hann window. So I should take that into account too for the weight.