Bump version to v1.0
It works, and I use it, let's put a number on it. I could call it 0.1,
and then release 0.2, and so on. But why not drop the eternal "0."
prefix, and just call it 1.0? It's just a number. It doesn't mean a lot
apart from going up on subsequent releases. Something like Semver hardly
applies here: it is an application, it does not expose an API, maybe
apart from the user interface. So let's just call it 1.0.